UpTask Project. Autentificacion finalizada. Agregado los endpoints para agregar miembros.

+++ uptask/backend/src/controllers/ProjectController.ts
@@ -1,22 +1,30 @@
 import {Request, Response} from "express"
 import Project from "../models/Project"
+import {Types} from "mongoose";
 
 export class ProjectController {
 
     static createProject = async (req: Request, res: Response)=> {
         try {
             const project = new Project(req.body)
-            await project.save()
+            project.manager = req.user._id
+            req.user.projects.push(new Types.ObjectId(project.id))
+            await Promise.allSettled([project.save(), req.user.save()])
             return res.status(200).send('Proyecto creado correctamente')
         }
         catch (error) {
+            console.log(error)
             return res.status(500).send('Error interno del servidor')
         }
     }
 
     static getAllProjects = async (req: Request, res: Response)=> {
         try {
-            const projects = await Project.find({}, undefined, undefined)
+            const projects = await Project.find({
+                $or: [
+                    {manager: {$in: req.user._id}}
+                ]
+            }, undefined, undefined)
             return res.json(projects)
         }
         catch (error) {
@@ -32,6 +40,10 @@ export class ProjectController {
                 const error = new Error('Projecto no encontrado')
                 return res.status(400).json({ error: error.message })
             }
+            if (project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Accion no valida')
+                return res.status(403).json({ error: error.message })
+            }
             return res.json(project)
         }
         catch (error) {
@@ -46,6 +58,10 @@ export class ProjectController {
             if (!project) {
                 const error = new Error('Projecto no encontrado')
                 return res.status(400).json({ error: error.message })
+            }
+            if (project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Accion no valida')
+                return res.status(403).json({ error: error.message })
             }
             project.projectName = req.body.projectName
             project.clientName = req.body.clientName
@@ -65,6 +81,10 @@ export class ProjectController {
             if (!project) {
                 const error = new Error('Projecto no encontrado')
                 return res.status(400).json({ error: error.message })
+            }
+            if (project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Accion no valida')
+                return res.status(403).json({ error: error.message })
             }
             await project.deleteOne()
             return res.send('Proyecto eliminado exitosamente.')
